starknet: move 'snip' to 'core' directory
