+++ packages/starknet/lib/src/core/index.dart
@@ -1,2 +1,3 @@
 export 'signer/index.dart';
+export 'snip/index.dart';
 export 'types/index.dart';
