+++ packages/starknet/lib/src/core/snip/index.dart
