+++ packages/starknet/lib/src/core/snip/snip12/index.dart
