+++ packages/starknet/lib/src/core/snip/snip12/typed_data/encode.dart
