+++ packages/starknet/lib/src/core/snip/snip9/outside_execution.freezed.dart
